requirements.txt
flask
flask-cors
python-dotenv
binance
pandas
numpy
